fix: fix index files
