+++ project/libs/comment-access/src/index.ts
@@ -1,5 +1,5 @@
-export { Commentary } from './lib/types/comment.interface';
-export { CommentAccessModule } from './lib/comment-access.module';
-export { CommentAccessRepository } from './lib/comment-access.repository';
-export { CommentAccessFactory } from './lib/comment-access.factory';
-export { CommentAccessEntity } from './lib/comment-access.entity';
+export * from './lib/types/comment.interface';
+export * from './lib/comment-access.module';
+export * from './lib/comment-access.repository';
+export * from './lib/comment-access.factory';
+export * from './lib/comment-access.entity';
