+++ project/libs/configuration/src/auth-service/utils.ts
@@ -1,0 +1,16 @@
+import { ConfigService } from '@nestjs/config';
+import { JwtModuleAsyncOptions } from '@nestjs/jwt';
+import { AUTH_ALIAS } from './constants';
+
+export function getJwtOptions(): JwtModuleAsyncOptions {
+  return {
+    useFactory: async (config: ConfigService) => {
+      return {
+        global: true,
+        secret: config.get<string>(`${AUTH_ALIAS}.jwtSecret`),
+        signOptions: { expiresIn: `${config.get<string>(`${AUTH_ALIAS}.expiresTokenIn`)}s` },
+      };
+    },
+    inject: [ConfigService],
+  };
+}
