+++ project/libs/configuration/src/database/mongodb/utils.ts
@@ -1,8 +1,6 @@
 import { MongooseModuleAsyncOptions } from '@nestjs/mongoose';
 import { ConfigService } from '@nestjs/config';
-import { JwtModuleAsyncOptions } from '@nestjs/jwt';
-import { MONGO_ALIAS } from './database/mongodb/constants';
-import { AUTH_ALIAS } from './auth-service/constants';
+import { MONGO_ALIAS } from './constants';
 
 function getMongoConnectionString({ username, password, host, port, databaseName, authDatabase }): string {
   return `mongodb://${username}:${password}@${host}:${port}/${databaseName}?authSource=${authDatabase}`;
@@ -20,19 +18,6 @@ export function getMongooseOptions(): MongooseModuleAsyncOptions {
           authDatabase: config.get<string>(`${MONGO_ALIAS}.authBase`),
           databaseName: config.get<string>(`${MONGO_ALIAS}.name`),
         }),
-      };
-    },
-    inject: [ConfigService],
-  };
-}
-
-export function getJwtOptions(): JwtModuleAsyncOptions {
-  return {
-    useFactory: async (config: ConfigService) => {
-      return {
-        global: true,
-        secret: config.get<string>(`${AUTH_ALIAS}.jwtSecret`),
-        signOptions: { expiresIn: `${config.get<string>(`${AUTH_ALIAS}.expiresTokenIn`)}s` },
       };
     },
     inject: [ConfigService],
