+++ project/libs/configuration/src/index.ts
@@ -1,5 +1,6 @@
-export * from './config.utils';
+export * from './auth-service/utils';
+export * from './database/mongodb/utils';
+export * from './file-service/constants';
 export { default as mongoRegister } from './database/mongodb/mongo.register';
 export { default as fileServiceRegister } from './file-service/file-sevice.register';
 export { default as authServiceRegister } from './auth-service/auth-sevice.register';
-export * from './file-service/constants';
