+++ project/libs/core/src/index.ts
@@ -1,4 +1,4 @@
-export { BaseEntity } from './lib/base.entity';
-export { BaseEntityInterface } from './lib/base-entity.interface';
-export { StorableEntity } from './lib/storable-entity.interface';
-export { EntityFactory } from './lib/entity-factory.interface';
+export * from './lib/base.entity';
+export * from './lib/base-entity.interface';
+export * from './lib/storable-entity.interface';
+export * from './lib/entity-factory.interface';
