+++ project/libs/file-access/src/index.ts
@@ -1,5 +1,5 @@
-export { FileAccessModule } from './lib/file-access.module';
-export { FileAccessRepository } from './lib/file-access.repository';
-export { FileAccessFactory } from './lib/file-access.factory';
-export { FileAccessEntity } from './lib/file-access.entity';
-export { UserFile } from './lib/types/file.interface';
+export * from './lib/file-access.module';
+export * from './lib/file-access.repository';
+export * from './lib/file-access.factory';
+export * from './lib/file-access.entity';
+export * from './lib/types/file.interface';
