+++ project/libs/like-access/src/index.ts
@@ -1,5 +1,5 @@
-export { UserLike } from './lib/types/like.interface';
-export { LikeAccessModule } from './lib/like-access.module';
-export { LikeAccessRepository } from './lib/like-access.repository';
-export { LikeAccessFactory } from './lib/like-access.factory';
-export { LikeAccessEntity } from './lib/like-access.entity';
+export * from './lib/types/like.interface';
+export * from './lib/like-access.module';
+export * from './lib/like-access.repository';
+export * from './lib/like-access.factory';
+export * from './lib/like-access.entity';
