+++ project/libs/post-access/src/index.ts
@@ -1,12 +1,12 @@
-export { UserPost } from './lib/types/base/base-post.interface';
-export { LinkPost } from './lib/types/link-post.interface';
-export { PhotoPost } from './lib/types/photo-post.interface';
-export { VideoPost } from './lib/types/video-post.interface';
-export { TextPost } from './lib/types/text-post.interface';
-export { CommonPost } from './lib/types/common-post.interface';
-export { QuotePost } from './lib/types/quote-post.interface';
-export { PostAccessModule } from './lib/post-access.module';
-export { PostAccessRepository } from './lib/post-access.repository';
-export { PostAccessFactory } from './lib/post-access.factory';
-export { PostAccessEntity } from './lib/post-access.entity';
+export * from './lib/types/base/base-post.interface';
+export * from './lib/types/link-post.interface';
+export * from './lib/types/photo-post.interface';
+export * from './lib/types/video-post.interface';
+export * from './lib/types/text-post.interface';
+export * from './lib/types/common-post.interface';
+export * from './lib/types/quote-post.interface';
+export * from './lib/post-access.module';
+export * from './lib/post-access.repository';
+export * from './lib/post-access.factory';
+export * from './lib/post-access.entity';
 export * from './lib/types/update-common-post.interface';
