+++ project/libs/subscription-access/src/index.ts
@@ -1,5 +1,5 @@
-export { Subscription } from './lib/types/subscription.interface';
-export { SubscriptionAccessModule } from './lib/subscription-access.module';
-export { SubscriptionAccessRepository } from './lib/subscription-access.repository';
-export { SubscriptionAccessFactory } from './lib/subscription-access.factory';
-export { SubscriptionAccessEntity } from './lib/subscription-access.entity';
+export * from './lib/types/subscription.interface';
+export * from './lib/subscription-access.module';
+export * from './lib/subscription-access.repository';
+export * from './lib/subscription-access.factory';
+export * from './lib/subscription-access.entity';
