+++ project/libs/user-access/src/index.ts
@@ -1,5 +1,5 @@
-export { UserAccessEntity } from './lib/user-access.entity';
-export { UserAccessRepository } from './lib/user-access.repository';
-export { UserAccessModule } from './lib/user-access.module';
-export { User } from './lib/types/user.interface';
-export { AuthUser } from './lib/types/auth-user.interface';
+export * from './lib/user-access.entity';
+export * from './lib/user-access.repository';
+export * from './lib/user-access.module';
+export * from './lib/types/user.interface';
+export * from './lib/types/auth-user.interface';
